Updated beginning-of-file comments: Contributions from N.Dattani were made since 2007 but weren't mentioned. AF3x3 program was written by N.Dattani but didn't have a copyright notice before (files written mainly by LeRoy did have that notice).
The software has been open source for a while now.

Original files from 2018 can be found from the earliest commits in Git,
and contributions afterwards are documented by the Git version control
system. Contributions before 2018 are fuzzy since we weren't using
version control, and because they were not listed properly in the
comments (see for example AF3x3 which had no authorship comments, but
the PotFit version is credited to N Dattani).